requirements.txt
discord.py==1.7.1
aiohttp==3.7.4.post0
requests==2.25.1